Compilation of everything now clean with -O0 and -Wall, and moldft with full optimization and -Wall.  Will work on making rest clean as time permits.

+++ src/apps/molresponse_v2/ResponseSolver.cpp
@@ -62,7 +62,7 @@ ResponseSolverPolicy<StaticRestrictedResponse>::CoupledResponseEquations(
 
   int i = 0;
   int si = 0;
-  for (int j = 0; j < num_orbitals; j++) {
+  for (size_t j = 0; j < num_orbitals; j++) {
     state_index.push_back(si);
     ii.push_back(i++);
   }
@@ -111,7 +111,7 @@ ResponseSolverPolicy<DynamicRestrictedResponse>::CoupledResponseEquations(
   std::vector<int> ii;
 
   int i = 0;
-  for (int j = 0; j < 2 * num_orbitals; j++) {
+  for (size_t j = 0; j < 2 * num_orbitals; j++) {
     state_index.push_back(0);
     ii.push_back(i++);
   }
